+++ .gitignore
@@ -40,3 +40,5 @@ yarn-error.log*
 next-env.d.ts
 
 *storybook.log
+
+certificates
